+++ src/plain.rs
@@ -132,11 +132,11 @@
 use core::convert::Infallible;
 
 use bitfield::bitfield;
-#[cfg(not(feature = "esp-dma"))]
+#[cfg(not(feature = "esp-hal-dma"))]
 use embedded_dma::ReadBuffer;
 use embedded_graphics::pixelcolor::RgbColor;
 use embedded_graphics::prelude::Point;
-#[cfg(feature = "esp-dma")]
+#[cfg(feature = "esp-hal-dma")]
 use esp_hal::dma::ReadBuffer;
 
 use super::Color;
@@ -488,7 +488,7 @@ impl<
     /// type FBType = DmaFrameBuffer<ROWS, COLS, NROWS, BITS, FRAME_COUNT>;
     /// let (_, tx_descriptors) = esp_hal::dma_descriptors!(0, FBType::dma_buffer_size_bytes());
     /// ```
-    #[cfg(feature = "esp-dma")]
+    #[cfg(feature = "esp-hal-dma")]
     pub const fn dma_buffer_size_bytes() -> usize {
         core::mem::size_of::<[Frame<ROWS, COLS, NROWS>; FRAME_COUNT]>()
     }
@@ -664,7 +664,7 @@ unsafe impl<
         const FRAME_COUNT: usize,
     > ReadBuffer for DmaFrameBuffer<ROWS, COLS, NROWS, BITS, FRAME_COUNT>
 {
-    #[cfg(not(feature = "esp-dma"))]
+    #[cfg(not(feature = "esp-hal-dma"))]
     type Word = u8;
 
     unsafe fn read_buffer(&self) -> (*const u8, usize) {
@@ -682,7 +682,7 @@ unsafe impl<
         const FRAME_COUNT: usize,
     > ReadBuffer for &mut DmaFrameBuffer<ROWS, COLS, NROWS, BITS, FRAME_COUNT>
 {
-    #[cfg(not(feature = "esp-dma"))]
+    #[cfg(not(feature = "esp-hal-dma"))]
     type Word = u8;
 
     unsafe fn read_buffer(&self) -> (*const u8, usize) {
@@ -1108,7 +1108,7 @@ mod tests {
     }
 
     #[test]
-    #[cfg(feature = "esp-dma")]
+    #[cfg(feature = "esp-hal-dma")]
     fn test_dma_framebuffer_dma_buffer_size() {
         let expected_size =
             core::mem::size_of::<[Frame<TEST_ROWS, TEST_COLS, TEST_NROWS>; TEST_FRAME_COUNT]>();