+++ CHANGELOG.md
@@ -15,17 +15,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   The new name avoids shadowing `embedded_graphics::DrawTarget::clear(Color)`.  
   Update your code: `fb.clear()` ➜ `fb.erase()`.  
   If you actually wanted the trait method, call `fb.clear(Color::BLACK)` instead.
+* Removed feature flags `esp32`, `esp32s3`, `esp32c6`.
+* Renamed feature `esp-dma` ➜ `esp-hal-dma`.
 
 ### Added
 
-- `skip-black-pixels` feature that gives a performance boot in some cases (#2)
+* `skip-black-pixels` feature that gives a performance boot in some cases (#2)
+* Removed feature flags `esp32`, `esp32s3`, `esp32c6`.
 
 ### Changed
-- almost double(!) performance of the set_pixel in the plain and latched `DmaFrameBuffers` (#2)
+
+* almost double(!) performance of the set_pixel in the plain and latched `DmaFrameBuffers` (#2)
+* Renamed feature `esp-dma` ➜ `esp-hal-dma`.
 
 ## [0.1.0] - 2025-06-14
 
-- initial version
+* initial version
 
 <!-- next-url -->
 [Unreleased]: https://github.com/liebman/hub75-framebuffer/compare/v0.1.0...HEAD